src/test/java/guru/springframeword/sfgrecipes/converters/UnitOfMeasureCommandToUnitOfMeasureTest.java
package guru.springframeword.sfgrecipes.converters;

import guru.springframeword.sfgrecipes.commands.UnitOfMeasureCommand;
import guru.springframeword.sfgrecipes.domain.UnitOfMeasure;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class UnitOfMeasureCommandToUnitOfMeasureTest {

    private static final Long LONG_VALUE = 1L;
    private static final String DESCRIPTION = "description";

    UnitOfMeasureCommandToUnitOfMeasure converter;

    @BeforeEach
    void setUp() {
        converter = new UnitOfMeasureCommandToUnitOfMeasure();
    }

    @Test
    void testNullParameter() {
        assertNull(converter.convert(null));
    }

    @Test
    void testEmptyObject() {
        assertNotNull(converter.convert(new UnitOfMeasureCommand()));
    }

    @Test
    void convert() {
        //given
        UnitOfMeasureCommand command = new UnitOfMeasureCommand()
                .setId(LONG_VALUE)
                .setDescription(DESCRIPTION);

        //when
        UnitOfMeasure uom = converter.convert(command);

        //then
        assertNotNull(uom);
        assertEquals(uom.getId(), LONG_VALUE);
        assertEquals(uom.getDescription(), DESCRIPTION);
    }
}